Move CreateOauthStateString function to client
