+++ internal/pkg/source/github/repo_auth_flow.go
@@ -11,15 +11,13 @@ package github
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"github.com/wso2/choreo-cli/internal/pkg/cmd/runtime"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"strconv"
 	"time"
+
+	"github.com/wso2/choreo-cli/internal/pkg/cmd/runtime"
 
 	"github.com/wso2/choreo-cli/internal/pkg/client"
 	"github.com/wso2/choreo-cli/internal/pkg/cmd/common"
@@ -31,12 +29,10 @@ func PerformGithubAuthorization(cliContext runtime.CliContext) bool {
 	consoleWriter := cliContext.Out()
 	getEnvConfig := config.CreateConfigReader(cliContext.EnvConfig(), client.EnvConfigs)
 
-	state, err := obtainState(cliContext)
-	if state == "" {
-		if err != nil {
-			common.PrintErrorMessage(consoleWriter, err.Error())
-		}
-		common.ExitWithErrorMessage(consoleWriter, "Error while initiating authorization flow")
+	//state, err := obtainState(cliContext)
+	state, err := cliContext.Client().CreateOauthStateString()
+	if err != nil {
+		common.ExitWithError(consoleWriter, "Could not initiate GitHub OAuth flow", err)
 	}
 
 	localServerPort := common.GetFirstOpenPort(localServerBasePort)
@@ -98,42 +94,4 @@ func shutdownServer(consoleWriter io.Writer, server *http.Server) {
 	if err != nil {
 		common.PrintError(consoleWriter, "Error shutting down the local server. Reason: ", err)
 	}
-}
-
-func obtainState(cliContext runtime.CliContext) (string, error) {
-	backendUrl := cliContext.EnvConfig().GetStringOrDefault(client.BackendUrl, client.EnvConfigs[client.BackendUrl])
-	accessToken := cliContext.UserConfig().GetStringOrDefault(client.AccessToken, client.UserConfigs[client.AccessToken])
-	req, err := client.NewRequest(backendUrl, accessToken, "GET", backendOauthStatePath, nil)
-	if err != nil {
-		return "", err
-	}
-
-	skipVerify, _ := strconv.ParseBool(cliContext.
-		EnvConfig().GetStringOrDefault(client.SkipVerify, client.EnvConfigs[client.SkipVerify]))
-	httpClient := client.NewClient(skipVerify)
-	resp, err := httpClient.Do(req)
-	if err != nil {
-		return "", err
-	}
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-	var stateObj struct {
-		State string `json:"state"`
-	}
-	if resp.StatusCode == http.StatusOK {
-		err := json.Unmarshal(body, &stateObj)
-		if err != nil {
-			return "", err
-		}
-	} else {
-		err = fmt.Errorf("error response received for state request: %s", string(body))
-		return "", err
-	}
-	err = resp.Body.Close()
-	if err != nil {
-		return "", err
-	}
-	return stateObj.State, nil
 }
